Don't update PagerTitleStrip titles when scrolled.
If the layout of a PagerTitleStrip tab is updated it forces an update of
the tab titles. If the PagerTitleStrip is scrolled during this event
it may no longer be centered over the currently active layout which
causes incorrect titles to be displayed.

There is no need for the labels to be forced during the time the user
is scrolling the tabs because the event handler for the scroll event also
updates the labels, as well as taking into account if the view has been
scrolled away from the active layout.

Change-Id: I34808462110bbf4a8fd09d3b58ef8138c7e0ef18